⭐ feat: 게시글 조회 구현 완료

+++ src/main/java/com/adoptpet/server/community/dto/TrendingArticleDto.java
@@ -11,5 +11,5 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 public class TrendingArticleDto {
     private Integer articleNo;
-    private Integer likeCnt;
+    private Long likeCnt;
 }
